+++ RELEASE-NOTES.txt
@@ -16,6 +16,7 @@
 * [*] Fix an issue with New Stats resetting the preferred preset after selecting a custom range [#24899]
 * [*] Show the Top 10 and 50 items in the Top List screen in a dedicated section [#24868]
 * [*] Add pull-to-refresh to the new Traffic tab [#24898]
+* [*] Improve the "Tags" picker in "Post Settings" [#24883]
 
 26.3.1
 ------
