+++ WordPress/Classes/ViewRelated/Post/Prepublishing/PrepublishingViewController.swift
@@ -319,14 +319,16 @@ final class PrepublishingViewController: UIViewController, UITableViewDataSource
 
     private func didTapTagCell() {
         let post = post as! Post
-        let tagPickerViewController = TagsViewController(blog: post.blog, selectedTags: post.tags) { [weak self] tags in
+        let viewModel = TagsViewModel(blog: post.blog, selectedTags: post.tags, mode: .selection(onSelectedTagsChanged: { [weak self] tags in
             guard let self else { return }
             WPAnalytics.track(.editorPostTagsChanged, properties: Constants.analyticsDefaultProperty)
 
             (self.post as! Post).tags = tags
             self.reloadData()
-        }
-        navigationController?.pushViewController(tagPickerViewController, animated: true)
+        }))
+        let view = PostTagsView(viewModel: viewModel)
+        let hostVC = UIHostingController(rootView: view)
+        navigationController?.pushViewController(hostVC, animated: true)
     }
 
     // MARK: - Categories (Post)
