+++ WordPress/Classes/ViewRelated/Tags/TagsViewModel.swift
@@ -2,6 +2,7 @@ import Foundation
 import WordPressKit
 import WordPressData
 import WordPressUI
+import WordPressShared
 
 typealias TagsPaginatedResponse = DataViewPaginatedResponse<RemotePostTag, Int>
 
@@ -94,11 +95,24 @@ class TagsViewModel: ObservableObject {
         }
     }
 
-    func search() async throws -> DataViewPaginatedResponse<RemotePostTag, Int> {
+    var isLocalSearchEnabled: Bool {
+        guard let response else { return false }
+        return !response.hasMore
+    }
+
+    func search() async throws -> TagsPaginatedResponse {
+        if let response, !response.hasMore {
+            let results = await StringRankedSearch(searchTerm: searchText)
+                .parallelSearch(in: response.items, input: \.name)
+            return try await TagsPaginatedResponse { _ in
+                TagsPaginatedResponse.Page(items: results, total: results.count, hasMore: false, nextPage: nil)
+            }
+        }
+
         let remoteTags = try await tagsService.searchTags(with: searchText)
 
-        return try await DataViewPaginatedResponse { _ in
-            return DataViewPaginatedResponse.Page(
+        return try await TagsPaginatedResponse { _ in
+            return TagsPaginatedResponse.Page(
                 items: remoteTags,
                 total: remoteTags.count,
                 hasMore: false,
@@ -117,6 +131,7 @@ class TagsViewModel: ObservableObject {
             selectedTagsSet.insert(lowercasedTagName)
             selectedTags.append(tagName)
         }
+        searchText = ""
     }
 
     func addNewTag(named name: String) {
